クライアントサイドのモジュール化
