+++ server.js
@@ -7,7 +7,6 @@ const http = require('http');
 const path = require('path');
 const utils = require('./src/modules/utils');
 const Game = require('./src/modules/game');
-const Player = require('./src/modules/player');
 // ステージごとのファイル読み込み
 const init = require('./src/modules/stage/server/init');
 const entry = require('./src/modules/stage/server/entry');
@@ -19,7 +18,7 @@ const app = express();
 const server = http.Server(app);
 const io = socketIO(server);
 
-// プレイヤーリスト
+// ゲームオブジェクト作成
 let game = new Game();
 // 接続が完了したときに呼び出す関数
 io.on('connection', function(socket) {
@@ -30,6 +29,7 @@ io.on('connection', function(socket) {
     // クライアントからstartがemitされた時
     socket.on('start', () => start.do(socket, game));
     // TODO: ここに追加していく
+
     // 通信終了時(ブラウザを閉じる/リロード/ページ移動)
     // TODO: つまりリロードすると復帰不可
     socket.on('disconnect', () => disconnect.do(io, socket, game));
