+++ src/client.js
@@ -3,25 +3,14 @@
 // import modules
 import {Init} from './modules/stage/client/init.js'
 import {Start} from './modules/stage/client/start.js'
-import {Utils} from './modules/stage/client/utils.js'
+import {MasterHandSelection} from './modules/stage/client/master_hand_selection.js'
+import {CannotPlay} from './modules/stage/client/cannot_play.js';
 // const init = require('./modules/stage/client/init');
 // Socket.IOを利用してサーバに接続
 const socket = io();
 const canvas = document.getElementById('canvas-2d');
 const context = canvas.getContext('2d');
 const startButton = document.getElementById('startButton');
-
-// サーバーから'state'がemitされたときの動作
-socket.on('state', (players) => {
-    clearDisplay();
-    // サーバーから送られてきたplayersをそれぞれcanvasに描画
-    Object.values(players).forEach((player) => {
-        // context.drawImage(playerImage, player.x, player.y);
-        // context.font = '20px Bold Arial';
-        // let id = player.id.toString()
-        // context.fillText('Player' + id, player.x, player.y - 20);
-    });
-});
 
 // サーバーから'chat_send_from_server'がemitされた時の動作(チャット用)
 socket.on('chat_send_from_server', function(data){appendMsg(data.value)});
@@ -45,78 +34,16 @@ function updateNumOfPepole(num) {
 // 接続完了('connect')時(最初の接続完了時に'connect'がサーバーからemitされる)
 socket.on('connect', function(){Init.do(socket);});
 // サーバーから'cannot_play'がemitされた時
-socket.on('cannot_play', cannotPlay);
+socket.on('cannot_play', function(){CannotPlay.do(context, canvas);});
 // サーバーから'start'がemitされた時(startステージ移行)
 socket.on('start', Start.do);
 // サーバーから'master_hand_selection'がemitされた時(master_hand_selectionステージ移行)
-socket.on('master_hand_selection', function(data){masterHandSelection(data);});
+socket.on('master_hand_selection', function(data){MasterHandSelection.do(data, context);});
 
 
-// エントリーフォームにsubmitされたときの動作
-$("#entryForm").submit(function(e){
-    var username = $("#userName").val();
-    // 名前入力フォーム非表示
-    document.getElementById("entryForm").style.display = 'none';
-    Utils.clearDisplay();
-    const fontSize = 20;
-    context.font = fontSize + 'px Bold Arial';
-    const message = '参加者が集まるまでしばらくお待ちください';
-    // メッセージをcanvas中央に配置
-    context.fillText(message, canvas.width / 2 - message.length * fontSize / 2, canvas.height / 2 - fontSize / 2);
-    // サーバーに'entry'を送信
-    socket.emit('entry', {username : username});
-    e.preventDefault(); // フォームによる/?への接続を止める(socketIDを一意に保つため)
-});
-
-// cannot_play画面
-function cannotPlay() {
-    // 名前入力フォーム表示
-    document.getElementById("entryForm").style.display = 'block';
-    Utils.clearDisplay();
-    const fontSize = 20;
-    context.font = fontSize + 'px Bold Arial';
-    const message = '現在プレイ中です しばらくお待ちください';
-    // メッセージをcanvas中央に配置
-    context.fillText(message, canvas.width / 2 - message.length * fontSize / 2, canvas.height / 2 - fontSize / 2);
-}
-
-// スタートボタンを押した時の動作
-startButton.onclick = function() {
-    startButton.style.display = 'none';
-    Utils.clearDisplay();
-    const fontSize = 20;
-    context.font = fontSize + 'px Bold Arial';
-    const message = '他の参加者がスタートするのを待っています...';
-    // メッセージをcanvas中央に配置
-    context.fillText(message, canvas.width / 2 - message.length * fontSize / 2, canvas.height / 2 - fontSize / 2);
-    socket.emit('start');
-}
-
-// master_hand_selection画面
-function masterHandSelection(data) {
-    Utils.clearDisplay();
-    // TODO: srcを変更しているのに，drawImageでの表示が変更されない
-    // TODO: 応急処置としてhtmlにサーバーにある画像を追加し，そこから持ってくることにしているがダサい
-    const y = 600;
-    data.player.hand._array.forEach((card, index) => {
-        var x = 400.0 / 3.0 + index * (100 + 400.0 / (3.0 * 5.0));
-        var image = document.getElementById(card.filename);
-        context.drawImage(image, 0, 0, image.width, image.height, x, y, 100, 150);
-    });
-    // プレイヤー名/スコアの表示
-    // 自分
-    const fontSize = 20;
-    context.font = fontSize + 'px Bold Arial';
-    context.fillText(data.player.name, 20, y);
-    context.fillText('スコア: ' + data.player.score, 30, y + 40);
-    // 他の人
-    console.log(data.others);
-    data.others.forEach((player, index) => {
-        console.log(player);
-        context.fillText(player.name, index * 400 + 200, 200);
-        context.fillText('スコア: ' + player.score, index * 400 + 200, 200 + 40);
-    });
-}
+/****************************
+ *  イベントハンドラの登録一覧
+ ****************************/
 
 // チャットフォームにsubmitされた時の動作
 $("#chatForm").submit(function(e){
@@ -125,3 +52,9 @@ $("#chatForm").submit(function(e){
     socket.emit("chat_send_from_client", {value : message});
     e.preventDefault(); // フォームによる/?への接続を止める
 });
+
+// エントリーフォームにsubmitされたときの動作
+$("#entryForm").submit(function(event){Init.entry(event, socket, context, canvas)});
+
+// スタートボタンを押した時の動作
+startButton.onclick = function() {Start.push(socket,context, canvas)};
