+++ src/modules/stage/client/start.js
@@ -9,4 +9,15 @@ export class Start {
         startButton.style.display = 'block';
         console.log('[debug] スタート待機状態');
     }
+
+    static push(socket, context, canvas) {
+        startButton.style.display = 'none';
+        Utils.clearDisplay();
+        const fontSize = 20;
+        context.font = fontSize + 'px Bold Arial';
+        const message = '他の参加者がスタートするのを待っています...';
+        // メッセージをcanvas中央に配置
+        context.fillText(message, canvas.width / 2 - message.length * fontSize / 2, canvas.height / 2 - fontSize / 2);
+        socket.emit('start');
+    }
 }
